Change price data type in product database entity

+++ TechZoneAPI/Data/Models/Product.cs
@@ -22,9 +22,9 @@ namespace TechZoneAPI.Data.Models
 
         public Brand Brand { get; set; }
 
-        public decimal Price { get; set; }
+        public int Price { get; set; }
 
-        public decimal? Delivery { get; set; }
+        public int? Delivery { get; set; }
 
         public string Description { get; set; }
